Added range option
using the -r (--rng) follwed by the issue range (3-8 for example) and followed by the url of the series (requires the -s <url>) will download the issue range selected.  

Example:

python3 comic-dl-ru.py -s https://readcomicsonline.ru/comic/the-walking-dead-2003 -r 170-180

will download the walking dead issues 170 through 180.

There is not yet support for issues that skip or change numbering scheme in the middle of the issues.